Dev

See merge request bms/scmp/scmp-api!360

+++ src/main/java/com/aiqin/bms/scmp/api/purchase/domain/PurchaseOrderDetails.java
@@ -82,15 +82,15 @@ public class PurchaseOrderDetails {
 
     @ApiModelProperty(value="预付款金额")
     @JsonProperty("advance_payment")
-    private Integer advancePayment;
+    private Long advancePayment;
 
     @ApiModelProperty(value="到付金额")
     @JsonProperty("amount_payable")
-    private Integer amountPayable;
+    private Long amountPayable;
 
     @ApiModelProperty(value="月结金额")
     @JsonProperty("month_amount")
-    private Integer monthAmount;
+    private Long monthAmount;
 
     @ApiModelProperty(value="到付付款期")
     @JsonProperty("payable_time")
