+++ src/main/java/com/aiqin/bms/scmp/api/purchase/service/PurchaseApplyService.java
@@ -31,7 +31,7 @@ public interface PurchaseApplyService {
 
     HttpResponse purchaseApplyStatus(PurchaseApply purchaseApply);
 
-    HttpResponse<PurchaseFlowPathResponse> applyProductDetail(Integer singleCount, Integer productPurchaseAmount, String skuCode,
+    HttpResponse<PurchaseFlowPathResponse> applyProductDetail(Integer singleCount, Long productPurchaseAmount, String skuCode,
                                                               String supplierCode, String transportCenterCode, Integer productCount);
 
     HttpResponse<PurchaseNewContrastResponse> purchaseContrast(PurchaseNewContrastRequest contrastRequest);
